+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- convenience.h useful functions, like options stringify
 
 ## 0.1.1 - 2017-08-22
 ### Added
@@ -28,5 +30,3 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## pre-0.1.0
 Free-style development. :)
-
-